+++ modules/pytket-quantinuum/tests/conftest.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 import os
-from typing import Tuple
+from typing import Any, Dict, List, Tuple
 
 import pytest
 from _pytest.fixtures import SubRequest
@@ -42,6 +42,94 @@ def mock_token() -> str:
     token_payload = {"exp": 3278815149.143694}
     mock_token = jwt.encode(token_payload, key="", algorithm="HS256")
     return mock_token
+
+
+@pytest.fixture()
+def mock_machine_info() -> Dict[str, Any]:
+    return {
+        "name": "H9-27",
+        "n_qubits": 12,
+        "gateset": [],
+        "n_classical_registers": 50,
+        "n_shots": 10000,
+        "system_family": "mock_family",
+        "system_type": "hardware",
+        "emulator": "H9-27E",
+        "syntax_checker": "H9-27SC",
+        "batching": True,
+        "wasm": True,
+    }
+
+
+@pytest.fixture()
+def sample_machine_infos() -> List[Dict[str, Any]]:
+    return [
+        {
+            "name": "H1-2SC",
+            "n_qubits": 12,
+            "gateset": ["RZZ", "Riswap", "Rxxyyzz"],
+            "n_classical_registers": 50,
+            "n_shots": 10000,
+            "system_type": "syntax checker",
+            "wasm": True,
+        },
+        {
+            "name": "H1-1SC",
+            "n_qubits": 12,
+            "gateset": ["RZZ", "Riswap", "Rxxyyzz"],
+            "n_classical_registers": 50,
+            "n_shots": 10000,
+            "system_type": "syntax checker",
+            "wasm": True,
+        },
+        {
+            "name": "H1-1E",
+            "n_qubits": 20,
+            "gateset": ["RZZ", "Riswap", "Rxxyyzz"],
+            "n_classical_registers": 50,
+            "n_shots": 10000,
+            "system_type": "emulator",
+            "batching": True,
+            "wasm": True,
+        },
+        {
+            "name": "H1-1",
+            "n_qubits": 20,
+            "gateset": ["RZZ", "Riswap", "Rxxyyzz"],
+            "n_classical_registers": 50,
+            "n_shots": 10000,
+            "system_family": "H1",
+            "system_type": "hardware",
+            "emulator": "H1-1E",
+            "syntax_checker": "H1-1SC",
+            "batching": True,
+            "wasm": True,
+        },
+        {
+            "name": "H1-2",
+            "n_qubits": 12,
+            "gateset": ["RZZ", "Riswap", "Rxxyyzz"],
+            "n_classical_registers": 50,
+            "n_shots": 10000,
+            "system_family": "H1",
+            "system_type": "hardware",
+            "emulator": "H1-2E",
+            "syntax_checker": "H1-2SC",
+            "batching": True,
+            "wasm": True,
+        },
+        {
+            "name": "H1-2E",
+            "n_qubits": 12,
+            "gateset": ["RZZ", "Riswap", "Rxxyyzz"],
+            "n_classical_registers": 50,
+            "n_shots": 10000,
+            "system_type": "emulator",
+            "batching": True,
+            "wasm": True,
+        },
+        {"name": "H1", "n_qubits": 20},
+    ]
 
 
 @pytest.fixture(name="mock_quum_api_handler", params=[True, False])
@@ -88,9 +176,21 @@ def fixture_mock_quum_api_handler(
     return api_handler
 
 
+@pytest.fixture(scope="module", name="authenticated_quum_handler")
+def fixture_authenticated_quum() -> QuantinuumAPI:
+    # Authenticated QuantinuumAPI used for the remote tests
+    # The credentials are taken from the env variables:
+    # PYTKET_REMOTE_QUANTINUUM_USERNAME and PYTKET_REMOTE_QUANTINUUM_PASSWORD
+    return QuantinuumAPI(  # type: ignore # pylint: disable=unexpected-keyword-arg
+        _QuantinuumAPI__user_name=os.getenv("PYTKET_REMOTE_QUANTINUUM_USERNAME"),
+        _QuantinuumAPI__pwd=os.getenv("PYTKET_REMOTE_QUANTINUUM_PASSWORD"),
+    )
+
+
 @pytest.fixture(name="authenticated_quum_backend")
 def fixture_authenticated_quum_backend(
     request: SubRequest,
+    authenticated_quum_handler: QuantinuumAPI,
 ) -> QuantinuumBackend:
     # Authenticated QuantinuumBackend used for the remote tests
     # The credentials are taken from the env variables:
@@ -103,14 +203,11 @@ def fixture_authenticated_quum_backend(
     # but other params can be specified when parametrizing the
     # authenticated_quum_backend
     if request.param is None:
-        backend = QuantinuumBackend("H1-1SC")
+        backend = QuantinuumBackend("H1-1SC", _api_handler=authenticated_quum_handler)
     else:
-        backend = QuantinuumBackend(**request.param)
-
-    backend._api_handler = QuantinuumAPI(  # type: ignore # pylint: disable=unexpected-keyword-arg
-        _QuantinuumAPI__user_name=os.getenv("PYTKET_REMOTE_QUANTINUUM_USERNAME"),
-        _QuantinuumAPI__pwd=os.getenv("PYTKET_REMOTE_QUANTINUUM_PASSWORD"),
-    )
+        backend = QuantinuumBackend(
+            _api_handler=authenticated_quum_handler, **request.param
+        )
     # In case machine_debug was specified by mistake in the params
     backend._MACHINE_DEBUG = False